Move G1Element into its own header file
