+++ src/openrct2/drawing/G1Element.h
@@ -1,0 +1,76 @@
+/*****************************************************************************
+ * Copyright (c) 2014-2026 OpenRCT2 developers
+ *
+ * For a complete list of all authors, please refer to contributors.md
+ * Interested in contributing? Visit https://github.com/OpenRCT2/OpenRCT2
+ *
+ * OpenRCT2 is licensed under the GNU General Public License version 3.
+ *****************************************************************************/
+
+#pragma once
+
+#include "../core/CallingConventions.h"
+#include "../core/FlagHolder.hpp"
+
+#include <vector>
+
+namespace OpenRCT2
+{
+    enum class G1Flag : uint8_t
+    {
+        hasTransparency, // Image data contains transparent pixels (0XFF) which will not be rendered
+        one,
+        hasRLECompression, // Image data is encoded using RCT2's form of run length encoding
+        isPalette,         // Image data is a sequence of palette entries R8G8B8
+        hasZoomSprite,     // Use a different sprite for higher zoom levels
+        noZoomDraw,        // Does not get drawn at higher zoom levels (only zoom 0)
+    };
+    using G1Flags = FlagHolder<uint16_t, G1Flag>;
+
+    struct G1Element
+    {
+        uint8_t* offset = nullptr; // 0x00
+        union
+        {
+            int16_t width = 0;  // 0x04
+            int16_t numColours; // If G1Flag::isPalette is set
+        };
+        int16_t height = 0; // 0x06
+        union
+        {
+            int16_t xOffset = 0; // 0x08
+            int16_t startIndex;  // If G1Flag::isPalette is set
+        };
+        int16_t yOffset = 0;      // 0x0A
+        G1Flags flags = {};       // 0x0C
+        int32_t zoomedOffset = 0; // 0x0E
+    };
+
+#pragma pack(push, 1)
+    struct G1Header
+    {
+        uint32_t numEntries = 0;
+        uint32_t totalSize = 0;
+    };
+    static_assert(sizeof(G1Header) == 8);
+#pragma pack(pop)
+
+    struct Gx
+    {
+        G1Header header;
+        std::vector<G1Element> elements;
+        std::unique_ptr<uint8_t[]> data;
+    };
+
+    struct StoredG1Element
+    {
+        uint32_t offset;       // 0x00 note: uint32_t always!
+        int16_t width;         // 0x04
+        int16_t height;        // 0x06
+        int16_t xOffset;       // 0x08
+        int16_t yOffset;       // 0x0A
+        G1Flags flags;         // 0x0C
+        uint16_t zoomedOffset; // 0x0E
+    };
+    static_assert(sizeof(StoredG1Element) == 0x10);
+} // namespace OpenRCT2
